refactor(cli): modularize command groups
