+++ src/modos/cli/__init__.py
@@ -1,0 +1,1 @@
+from modos.cli.main import cli as cli
